refactor(team): extract requests/interviews into broker_requests_interviews.go
Slice 11. Moves the human-interview / approval-request HTTP surface
and helpers (~752 lines) into broker_requests_interviews.go:
- Helpers: requestIsActive, requestBlocksMessages, requestIsHumanInterview,
  requestNeedsHumanDecision, requestOptionDefaults, enrichRequestOptions,
  normalizeRequestOptions, findRequestOption, formatRequestAnswerMessage,
  activeRequests, firstBlockingRequest, firstActiveHumanInterview,
  humanSenderMayCancelInterviews, normalizeRequestKind, normalizeRequestOptionID
- Locked methods: cancelRequestLocked, cancelActiveHumanInterviewsLocked,
  unblockTasksForAnsweredRequestLocked
- HTTP handlers: handleRequests, handleGetRequests, handlePostRequest,
  handleRequestAnswer, handleGetRequestAnswer, handlePostRequestAnswer,
  handleInterview, handlePostInterview, handleGetInterview,
  handleInterviewAnswer, handleGetInterviewAnswer, handlePostInterviewAnswer
- reqAnswerSummary

Relocates 8 request/interview tests from broker_test.go into
broker_requests_interviews_test.go (lifecycle, cross-channel scope,
cancel-on-human, decision blocking defaults, custom-text option).

Adds three characterization tests:
- TestNormalizeRequestOptions_StableOrderAndIDs
- TestEnrichRequestOptions_AddsKindDefaults
- TestCancelActiveHumanInterviewsLocked_NoOpWhenNonePending

Total package coverage: 63.1% (vs 62.9% baseline floor).
